refactor: parser
