+++ driving_log_replayer_analyzer/driving_log_replayer_analyzer/analysis/obstacle_segmentation.py
@@ -15,7 +15,7 @@
 from pathlib import Path
 
 from driving_log_replayer_analyzer.config.obstacle_segmentation import load_config
-from driving_log_replayer_analyzer.data.jsonl_parser import JsonlParser
+from driving_log_replayer_analyzer.data.obstacle_segmentation import JsonlParser
 from driving_log_replayer_analyzer.data.obstacle_segmentation import fail_3_times_in_a_row
 from driving_log_replayer_analyzer.plot.bird_view_plot import BirdViewPlot
 from driving_log_replayer_analyzer.plot.line_plot import LinePlot
